formatter and deleted unwanted imports

+++ src/proposal/base/ModuleManager.sol
@@ -74,7 +74,7 @@ abstract contract ModuleManager is
     // Constants
 
     /// @dev Marks the maximum amount of Modules a Proposal can have to avoid out-of-gas risk.
-    uint256 private constant MAX_MODULE_AMOUNT = 128;
+    uint private constant MAX_MODULE_AMOUNT = 128;
 
     //--------------------------------------------------------------------------
     // Storage
